Cleanup the merge of OpenAPI and API-index views and tests
There were 2 test_openapi files and one had the api-root tests in them
too. The "APIIndexView" is also a base class for many other dataset
types, which really deserved to be in a separate file too.
