+++ src/dso_api/dynamic_api/views/__init__.py
@@ -1,5 +1,6 @@
 """All views for the dynamically generated API, split by protocol type."""
-from .api import APIIndexView, DynamicApiViewSet, reload_patterns, viewset_factory
+from .api import DynamicApiViewSet, reload_patterns, viewset_factory
+from .index import APIIndexView
 from .mvt import DatasetMVTIndexView, DatasetMVTSingleView, DatasetMVTView
 from .oauth import oauth2_redirect
 from .wfs import DatasetWFSIndexView, DatasetWFSView
